+++ lib/RPS_Run_game.rb
@@ -1,0 +1,7 @@
+require_relative 'RPS_Game'
+
+players = Game.new()
+players.run_full_game()
+
+
+
